Wrote simple factory tests
